Remove distracting info from documentation
Listing three separate ways to parse command-line arguments in Rust is
not really helpful. If you are using `cargo metadata`, it's quite
likely that you are already a somewhat of an expert in Rust, and it's
rather unlikely that you literally want to just wrap it as is.

OTOH, the fact that most of the crate-level docs is occupied by
argument parsing makes finding the relevant info more difficult.