tests/selftest.rs: chdir
These tests are invoked in the source directory.  In an out of tree
build, that may be totally different to the invocation directory.

We really want to chdir back to the cargo invocation directory.
Otherwise we risk (for example) picking up inappropriate cargo
configuration somewhere in the parents of the source directory.

This is not possible because cargo doesn't tell us what it is.

For now, the best guess is to change to CARGO_HOME.  That means
at least that we will pick up the right cargo configuration.

A consequence is that we must no longer use current_dir(), but rather
the directory containing the manifest, which cargo supplies in
CARGO_MANIFEST_DIR.