Add MetadataCommand.parse
This makes the useful parsing functionality available to callers who
don't want to process output generated earlier, for any reaoson.

It's still not possible to completely decouple cargo_command and
parse, and just use them separately, because we embed
--format-version.

Providing this public parse() method is sensible because we ought not
to expect our caller to call serde_json themselves.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>