Add support for advanced feature configuration
The goal is to emulate

  cargo build --no-default-features --features foo

which cannot be expressed in cargo_metadata today. As of this commit,
users may call `features()` multiple times, which enables them to
express the above. The test suite in the previous commit now passes.

This change should be backwards compatible. However, anyone who is
relying on subsequent calls to `features()` that squash previous
calls are now broken.